% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{conf_macro_generator}
\alias{conf_macro_generator}
\title{General Confluence macro builder for internal use}
\usage{
conf_macro_generator(
  type = c("inline", "block"),
  name,
  parameters = NULL,
  body = NULL
)
}
\arguments{
\item{type}{inline or block code style to be used for the HTML content}

\item{name}{Confluence macro name}

\item{parameters}{named list of optional macro parameters}

\item{body}{optional \code{confl-ac-rich-text-body} content}
}
\value{
HTML
}
\description{
General Confluence macro builder for internal use
}
\keyword{internal}
