% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{confl_macro_excerpt}
\alias{confl_macro_excerpt}
\title{Generate Confluence macro for an excerpt block}
\usage{
confl_macro_excerpt(hidden = TRUE, body)
}
\arguments{
\item{hidden}{if the \code{body} should be shown on the actual page}

\item{body}{HTML content of the excerpt}
}
\value{
HTML as string
}
\description{
Generate Confluence macro for an excerpt block
}
\references{
\url{https://confluence.atlassian.com/doc/excerpt-macro-148062.html}
}
