% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{confl_macro_info}
\alias{confl_macro_info}
\alias{confl_macro_tip}
\alias{confl_macro_note}
\alias{confl_macro_warning}
\title{Generate Confluence macro for a Info, Tip, Note, or Warning block}
\usage{
confl_macro_info(body)

confl_macro_tip(body)

confl_macro_note(body)

confl_macro_warning(body)
}
\arguments{
\item{body}{HTML content of the block}
}
\value{
HTML as string
}
\description{
Generate Confluence macro for a Info, Tip, Note, or Warning block
}
\references{
\url{https://confluence.atlassian.com/doc/info-tip-note-and-warning-macros-51872369.html}
}
