% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{confl_macro_toc}
\alias{confl_macro_toc}
\title{Generate Confluence macro for dynamic Table of Contents}
\usage{
confl_macro_toc(levels)
}
\arguments{
\item{levels}{max number of levels to show}
}
\value{
HTML as string
}
\description{
Generate Confluence macro for dynamic Table of Contents
}
\examples{
confl_macro_toc(2)
}
