% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflr-package.R
\docType{package}
\name{conflr}
\alias{conflr}
\alias{conflr-package}
\title{R Client for 'Confluence' API}
\description{
Provides utilities for working with various 'Confluence' API 
    <https://docs.atlassian.com/ConfluenceServer/rest/latest/>, including a
    functionality to convert an R Markdown document to 'Confluence' format and
    upload it to 'Confluence' automatically.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://line.github.io/conflr/}
  \item \url{https://github.com/line/conflr}
  \item Report bugs at \url{https://github.com/line/conflr/issues}
}

}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{hiroaki.yutani@linecorp.com}

Other contributors:
\itemize{
  \item  LINE Corporation [copyright holder]
}

}
