% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{confluence_expand}
\alias{confluence_expand}
\title{Generate Confluence macro for an expand block}
\usage{
confluence_expand(title, content)
}
\arguments{
\item{title}{defines the text that appears next to the expand/collapse icon}

\item{content}{this HTML content will be visible when someone clicks the macro title}
}
\value{
HTML as string
}
\description{
Generate Confluence macro for an expand block
}
\note{
\code{content} needs to be HTML, so look at \code{commonmark::markdown_html}, \code{pander::pander} and eg \code{xtable} for doing the conversion before passing to \code{confluence_expand}
}
\examples{
\dontrun{
confluence_expand(
  'Example block',
  commonmark::markdown_html(pander::pander_return(list(a = list(b = 4), c = 2))))
}
}
\references{
\url{https://confluence.atlassian.com/doc/expand-macro-223222352.html}
}
