% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{confl_create_post_from_Rmd}
\alias{confl_create_post_from_Rmd}
\title{Publish R Markdown Document to 'Confluence'}
\usage{
confl_create_post_from_Rmd(
  Rmd_file,
  interactive = NULL,
  params = NULL,
  ...,
  title = NULL,
  space_key = NULL,
  type = NULL,
  parent_id = NULL,
  update = NULL,
  use_original_size = FALSE
)
}
\arguments{
\item{Rmd_file}{Path to a .Rmd file.}

\item{interactive}{If \code{FALSE}, shiny interface is not launched.}

\item{params}{If provided, a list of named parameters that override custom
params in the YAML front-matter.}

\item{...}{Ignored.}

\item{title}{Title of the post.}

\item{space_key}{The space key to find content under.}

\item{type}{If provided, this overwrites the YAML front matter type}

\item{parent_id}{The page ID of the parent pages.}

\item{update}{If \code{TRUE}, overwrite the existing page (if it exists).}

\item{use_original_size}{If \code{TRUE}, use the original image sizes.}
}
\description{
Knit and post a given R Markdown file to 'Confluence'.
}
\details{
\code{title}, \code{type}, \code{space_key}, \code{parent_id}, \code{update}, and \code{use_original_size}
can be specified as \code{confluence_settings} item in the front-matter of the
Rmd file to knit. The arguments of \code{confl_create_post_from_Rmd()} overwrite
these settings if provided.
}
