% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{confl_create_post_from_Rmd}
\alias{confl_create_post_from_Rmd}
\title{Publish R Markdown Document to 'Confluence'}
\usage{
confl_create_post_from_Rmd(
  Rmd_file,
  interactive = NULL,
  params = NULL,
  ...,
  title = NULL,
  space_key = NULL,
  type = NULL,
  parent_id = NULL,
  update = NULL,
  use_original_size = FALSE
)
}
\arguments{
\item{Rmd_file}{Path to a .Rmd file.}

\item{interactive}{If \code{FALSE}, shiny interface is not launched.}

\item{params}{If provided, a list of named parameters that override custom
params in the YAML front-matter.}

\item{...}{Ignored.}

\item{title}{If provided, this overwrites the YAML front matter title.}

\item{space_key}{If provided, this overwrites the YAML front matter space_key.}

\item{type}{If provided, this overwrites the YAML front matter type}

\item{parent_id}{If provided, this overwrites the YAML front matter parent_id}

\item{update}{If \code{TRUE}, overwrite the existing page (if it exists).}

\item{use_original_size}{If \code{TRUE}, use the original image sizes.}
}
\description{
Knit and post a given R Markdown file to 'Confluence'.
}
