% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{confl_macro_jira}
\alias{confl_macro_jira}
\title{Generate Confluence macro referencing a Jira ticket}
\usage{
confl_macro_jira(id)
}
\arguments{
\item{id}{Jira ticket id, eg CONFLR-XXXX}
}
\value{
HTML as string
}
\description{
Generate Confluence macro referencing a Jira ticket
}
\examples{
confl_macro_jira('CONFLR-42')
}
