% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflr-package.R
\docType{package}
\name{conflr}
\alias{conflr}
\alias{conflr-package}
\title{R Client for 'Confluence' API}
\description{
Post R Markdown documents as 'Confluence' pages via its API.
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{hiroaki.yutani@linecorp.com}

Other contributors:
\itemize{
  \item  LINE Corporation [copyright holder]
}

}
